requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
fastapi==0.104.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
openai==1.23.6
peewee==3.17.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
twilio==8.10.3
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.4

